requirements.txt
selene==2.0.0rc8
pytest==8.1.1
allure-pytest==2.13.5
requests==2.31.0
appium-python-client==4.0.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settinga==2.2.1